pyproject.toml
[project]
name = "citation-tracker"
version = "0.1.0"
description = "Track citations of academic papers and analyse their arguments"
requires-python = ">=3.12"
dependencies = [
    "httpx",
    "pymupdf4llm",
    "openai",
    "resend",
    "click",
    "pyyaml",
    "python-dotenv",
    "tenacity",
    "rich",
]

[project.scripts]
citation-tracker = "citation_tracker.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/citation_tracker"]
